# api/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Village Homepage API
  version: 1.0.0-alpha
  description: |
    RESTful API for Village Homepage platform covering authentication, user preferences,
    feature flags, rate limiting, and widget data.

    ## Authentication
    - **OAuth2**: Google, Facebook, Apple authentication flows
    - **Anonymous Cookies**: Anonymous users get `vu_anon_id` cookie for preference storage
    - **JWT Tokens**: Admin endpoints require JWT bearer tokens with `super_admin` role

    ## Rate Limiting
    All API endpoints enforce tier-based rate limiting. Rate limit information is returned
    in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed in window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Window`: Window duration in seconds

    ## Schema Versioning
    DTOs include `schema_version` fields for forward-compatible migrations. Services handle
    older versions gracefully by applying defaults for missing fields.

    ## Policy References
    - **P1 (GDPR/CCPA)**: User data is mergeable and deletable
    - **P9 (Anonymous Cookie Security)**: Anonymous users can store preferences pre-auth
    - **P14 (Rate Limiting)**: All endpoints enforce tier-based limits

  contact:
    name: VillageCompute Team
    url: https://villagecompute.com
servers:
  - url: http://localhost:8080
    description: Development
  - url: https://homepage-beta.villagecompute.com
    description: Beta
...
security:
  - cookieAuth: []
  - oauth2: []
...
paths:
...
  /api/widgets/news:
    get:
      summary: Get personalized news feed widget data
      description: |
        Returns personalized news feed items based on user's RSS source subscriptions and
        AI-tagged content. Supports pagination and filtering by category/topics.

        **Personalization:** Items are filtered by:
        - User's active RSS source subscriptions
        - User's topic preferences (matched against AI tags)
        - Optional category and topics query parameters

        **Caching:** 5-minute TTL (Cache-Control: max-age=300, must-revalidate)

        **Rate Limiting:** 100 reads/hour for logged_in tier
      tags:
        - Widgets
      parameters:
        - name: limit
          in: query
          description: Maximum items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: category
          in: query
          description: Filter by RSS source category
          schema:
            type: string
        - name: topics
          in: query
          description: Comma-separated topics for AI tag matching
          schema:
            type: string
          example: "technology,ai,science"
      responses:
        '200':
          description: Personalized news feed data
          headers:
            Cache-Control:
              description: Caching directives
              schema:
                type: string
              example: "max-age=300, must-revalidate"
            Vary:
              description: Response varies by Authorization header
              schema:
                type: string
              example: "Authorization"
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Window:
              $ref: '#/components/headers/X-RateLimit-Window'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsWidgetType'
              example:
                items:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    source_id: "660e8400-e29b-41d4-a716-446655440000"
                    title: "Latest AI Breakthrough in Machine Learning"
                    url: "https://example.com/article"
                    description: "Researchers announce new transformer architecture..."
                    author: "Jane Doe"
                    published_at: "2025-01-09T10:00:00Z"
                    ai_tags:
                      topics: ["ai", "machine-learning", "research"]
                      sentiment: "positive"
                      categories: ["Technology", "Science"]
                      confidence: 0.95
                    ai_tagged: true
                    fetched_at: "2025-01-09T10:30:00Z"
                total_count: 150
                limit: 20
                offset: 0
                metadata:
                  click_tracking_enabled: true
                  click_tracking_base_url: "/track/click"
                  feature_flags:
                    ai_tags_enabled: true
                    personalization_enabled: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    NewsWidgetType:
      type: object
      required:
        - items
        - total_count
        - limit
        - offset
        - metadata
      properties:
        items:
          type: array
          description: List of feed items for current page
          items:
            $ref: '#/components/schemas/FeedItemType'
        total_count:
          type: integer
          description: Total number of items available (before pagination)
          example: 150
        limit:
          type: integer
          description: Maximum items per page
          example: 20
        offset:
          type: integer
          description: Pagination offset
          example: 0
        metadata:
          type: object
          description: Additional context (feature flags, click tracking config)
          properties:
            click_tracking_enabled:
              type: boolean
              example: true
            click_tracking_base_url:
              type: string
              example: "/track/click"
            feature_flags:
              type: object
              additionalProperties: true
              example:
                ai_tags_enabled: true
                personalization_enabled: true
    FeedItemType:
      type: object
      required:
        - id
        - source_id
        - title
        - url
        - published_at
        - ai_tagged
        - fetched_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the feed item
          example: "550e8400-e29b-41d4-a716-446655440000"
        source_id:
          type: string
          format: uuid
          description: RSS source UUID
          example: "660e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: Article title
          example: "Latest AI Breakthrough in Machine Learning"
        url:
          type: string
          format: uri
          description: Article URL
          example: "https://example.com/article"
        description:
          type: string
          nullable: true
          description: Article summary/description
          example: "Researchers announce new transformer architecture..."
        author:
          type: string
          nullable: true
          description: Article author
          example: "Jane Doe"
        published_at:
          type: string
          format: date-time
          description: Publication timestamp (ISO-8601)
          example: "2025-01-09T10:00:00Z"
        ai_tags:
          $ref: '#/components/schemas/AiTagsType'
          nullable: true
          description: AI-generated tags (topics, sentiment, categories)
        ai_tagged:
          type: boolean
          description: Whether AI tagging has completed for this item
          example: true
        fetched_at:
          type: string
          format: date-time
          description: Timestamp when item was fetched from RSS source (ISO-8601)
          example: "2025-01-09T10:30:00Z"
    AiTagsType:
      type: object
      required:
        - topics
        - sentiment
        - categories
        - confidence
      properties:
        topics:
          type: array
          items:
            type: string
          description: AI-identified topics/keywords from article content
          example: ["ai", "machine-learning", "research"]
        sentiment:
          type: string
          description: Sentiment analysis result
          enum:
            - positive
            - negative
            - neutral
            - mixed
          example: "positive"
        categories:
          type: array
          items:
            type: string
          description: Content categories (Technology, Business, Science, etc.)
          example: ["Technology", "Science"]
        confidence:
          type: number
          format: double
          minimum: 0.0
          maximum: 1.0
          description: AI confidence score (0.0-1.0) for tag quality
          example: 0.95
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Invalid request parameters
  responses:
    BadRequest:
      description: Invalid request parameters or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Validation failed: schema_version must be at least 1"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication required"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Window:
          $ref: '#/components/headers/X-RateLimit-Window'
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
          example: 45
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Rate limit exceeded. Retry after 45 seconds."
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "An unexpected error occurred"
  headers:
    X-RateLimit-Limit:
      description: Maximum requests allowed in window
      schema:
        type: integer
      example: 60
    X-RateLimit-Remaining:
      description: Requests remaining in current window
      schema:
        type: integer
      example: 45
    X-RateLimit-Window:
      description: Window duration in seconds
      schema:
        type: integer
      example: 60
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        OAuth2 authentication supporting Google, Facebook, and Apple providers.
        Anonymous users receive `vu_anon_id` cookie for preference storage before authentication.
      flows:
        authorizationCode:
          authorizationUrl: /auth/oauth/{provider}
          tokenUrl: /auth/token
          scopes: {}

    cookieAuth:
      type: apiKey
      in: cookie
      name: vu_anon_id
      description: |
        Anonymous user cookie for preference storage before authentication.
        Automatically issued on first visit and merged with user account on login.

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token for admin endpoints. Requires `super_admin` role claim.
        Obtained via OAuth2 authentication flow for privileged users.